qcacld-3.0: Cast to int32_t before comparing two different enum variables
To avoid build failure when using '-Werror' and
'-Wenum-compare' together in compiler options(which
will treat any warnings about comparisons between
different enumeration types as errors), cast both to
int32_t before comparing.

CRs-Fixed: 3984502

+++ core/wma/src/wma_main.c
@@ -6418,7 +6418,7 @@ static void wma_update_mlme_aux_dev_caps(struct wlan_objmgr_psoc *psoc,
 	struct wlan_mlme_aux_dev_caps
 		wlan_mlme_aux0_dev_caps[WLAN_MLME_HW_MODE_MAX] = {0};
 
-	if (WMI_HOST_HW_MODE_MAX != WLAN_MLME_HW_MODE_MAX)
+	if ((int32_t)WMI_HOST_HW_MODE_MAX != (int32_t)WLAN_MLME_HW_MODE_MAX)
 		wma_err("struct define mismatch, pls fix it.");
 
 	num_aux_dev_caps =
